render mention inline
